requirements.txt
Flask==2.3.6
google-cloud-firestore==2.21.0
google-cloud-storage==2.21.0
firebase-admin==6.1.0
python-dotenv==1.1.0
gunicorn==21.2.0
